ALL TESTS PASS except newAvailTest().
-Reworked L2 chunk expiration mechanism.  Invalid chunks eventually get cleaned up, and it will be possible to add the ability to limit memory consumed by level two chunks and/or the maximum number of chunks.  The primary changes are:
  -Code and continuations refer to chunks by reference.
  -A list of WeakReferences and a ReferenceQueue are used to recyclable chunk indices.  The WeakReference subclass instances track the contingent implementation sets for a chunk.
  -Implementation sets keep track of the indices of the chunks that depend on them.

-Tuple creation uses a static mutableObjectOfSize(n) instead of the clumsy combination of the static isMutableSize(bool, n) and create(n).

-Lots of lots of little cleanups to code and comments.

+++ Avail/src/com/avail/interpreter/Interpreter.java
@@ -247,7 +247,7 @@ public abstract class Interpreter
 		final AvailObject methodName,
 		final AvailObject method)
 	{
-		final short numArgs = method.type().numArgs();
+		final int numArgs = method.type().numArgs();
 		final AvailObject returnsBlock =
 			ClosureDescriptor.createStubForNumArgsConstantResult(
 				numArgs,
